ex06/src/llm_factory.py
"""ex06 — LLM 인스턴스 팩토리."""

import os

from dotenv import load_dotenv

load_dotenv()


def build_llm(temperature=0.0):
    """환경 변수에 따라 LLM 인스턴스를 생성한다."""
    provider = os.getenv("LLM_PROVIDER", "ollama").lower()

    if provider == "openai":
        from langchain_openai import ChatOpenAI
        return ChatOpenAI(
            model=os.getenv("OPENAI_MODEL", "gpt-4o-mini"),
            temperature=temperature,
            api_key=os.getenv("OPENAI_API_KEY"),
        )

    # 기본값: Ollama
    from langchain_ollama import ChatOllama
    return ChatOllama(
        model=os.getenv("OLLAMA_MODEL", "llama3.1:8b"),
        base_url=os.getenv("OLLAMA_BASE_URL", "http://localhost:11434"),
        temperature=temperature,
    )
